fixed non-standard urls all over nmis which conflicted with url-encoding (=percent encoding)
urls were like somewhere?%name=val&... but have to be
somewhere?name=val, without the % because that triggers
percent-encoding. the mess worked so far and mostly, because the
common first arg is "conf", and "co" is not a valid hex number, so
browsers ignored the dud percent encoding. but as soon as you use
"act=bla" (eg. in find.pl), you get "ac" which is valid hex and makes
the browser replace the %ac with latin-1 symbol for 'negation' - and
the handling script receives badly broken url arguments.