final bit of work for SUPPORT-199
all admin and support scripts now work without relying on the removed
database cache structure in the xzy-node files.

the gui components now allow the selection of non-collected interfaces
for displaying of graphs.

the graphtype clearing-and-recreation in the xyz-node files which
happened on collect was modified (no more blanked delete), so that
nmis doesn't forget whether pkts_hc is available for non-collected
interfaces - this is needed for the 'interface details' page.

loadtable and loadnodeinfotable gained an optional argument
suppress_errors for quieter handling of speculative reads that are not
certain to succeed.