snmp process state collection for servers is now unconditional, and the process data is stored in xyz-node.nmis. note: this automatically enables process memsize  thresholds.
fixed a small bug that showed the 'service list' tab even if no services
are collected/available.

fixed a bug in some less-commonly-used cgi scripts which ran createHRButtons()
without a live sys object - which fails because createhrbuttons now needs
the system methods to query the node's nodeinfo data for service checks
(before that used the removed database cache, now it looks for instances
of a service graph).