OMK-678, limit run time of nmis processes
added max_child_runtime config option (= seconds to live)
added code to enforce that for type=collect and type=update operations.
if mthread is on, each child is given that limit independently. otherwise,
the one single iterating process is limited.

if the ttl expires, an "nmis runtime exceeded" event is created and the process
dies with logmsg+stderr output.