fix for NMIS-44, split service cpu and mem graphs
new graphs use rrdtool's unit scaling for mem; for cpu it's only used for the
graph, not the value printing.

also revised the (now deprecated) cpumem graph to get rid of
the centi-seconds.

rrddraw.pl was updated to HIDE cpu and memory on the generic service graph
(would show these with bad units since the 'print extra values' change
a few weeks ago)