fixed a number of installer bugs/shortcomings
func.pm now fixes a dir's perms if setfileprotdirectory is used,
NOT just the perms of things within the dir.

nmis.pl now creates the var/nmis_system dir on the go if it isn't
present. the installer should take care of that, but there's a
chicken-egg issue when nmis is locked during installation where we
look bad.

installer now copies the tables.nmis and table-x.nmis files from install
to conf, unconditionally. also copies the new tables
events.nmis, businessservices, customers and servicestatus if not present.
install.log is now correctly created/updated in the install dir, not just
a cutoff part of it as before. less interaction needed, source dir is
autodetected  and destdir is fixed (but the site=X option for custom install
is shown to user)