workaround proc::processtable issue in func::find_nmis_processes
symptom: max nmis processes in runtime graph is never higher than 1.

some versions of p::p (e.g. the one on magni) get the 'fname' wrong if the process has
changed its $0 (like most of nmis does). not sure it's caused by the combo of p::p and kernel.

fix: detect this case and get the actual exename in that case