OMK-3232 allow selftest state reset from the gui - now works with nmis js refresh horrors
the selftest reset function now reloads the whole gui if widgetted; otherwise the commonv8 js
picks up the wrong url and attaches that to the widget (which then automatically performs the reset
on every page refresh!). in non-widgetted mode the  redirect is good enough.

also added the reset button to the view selftest page iff errors are present