fix for OMK-3239 and OMK-2965, i think: make snmp back off and retry when a message size exceeded snmp error is encountered
new code reduces the maxrepetitions by 25%, repeats that up to four times;
new value then persists for the remainder of the session lifetime.
this is debug logged only for now: fixme may want to logMsg?

updates nmis.pl to use the fast-good-column-bulk code for service data polling
but FIXME have not removed the by-row-slow-snail code yet; should be done
before full release.

a misconfigured cupid with maxrepetitions 500 now works; the first one or
two index ops usually fail (because of the limited number of retries), then
things work ok for the remainder of the run.