OMK-1887, improve nagios support for multi-valued 'performance data' responses
related to OMK-3479, which only partially fixed the mess the nsclient++ produces.

now: all numbers of perfdata responses are extracted,
if the unit is recognized, the value is scaled for rrd - NOT for servicestatus, where it
remains as-is but sans unit.

all perfdata labels are cleaned up for rrd, anything outside of [a-z0-9_] is replaced by an _.
again NO change for servicestatus, where this isn't a problem.