OMK-4249, nmis8: be smarter about conflicting processes with fast polls
we have no markers for attempted/in progress things in nmis8 (nodefile cannot
be rewritten that often).

result: collect job running isn't contributing to last-poll-x info and the node
is misconsidered to be a candidate on the next run (just 60s later...)

fix: in that case, synthesise the last-poll-x markers from the offending process
start time instead. that way, process start time + desired frequency still provides
the required safety net.