OMK-4341, cleanup/followup for cd51598; node.pl export to csv not working for some graphs
issue was code in getrrdstats and getrrdashash that didn't synthesise an $item from $index,
only the other way round. common-database entries with $item and not $index would thus silently not
resolve to a db filename.

keith replaced (but just commented) that gettypename->getfilename sequence with a call to
the sys::getdbname wrapper, which does call gettypename->getfilename behind the scenes...but after
synthesising item/index both ways.