OMK-4710, nmis8: handle dead/comatose fpingd and dud fpingd results better
fpingd is now (re)started if none is alive or if the fpingd output file is older than config daemon_fping_maxage
(or default 15 minutes).

fpingd results are now treated more precisely, ie. for each node independently.
if the data is missing for this node or stale (same criteria as above) then an internal ping is attempted.
if that fallback was made, then node up and down events are handled by nmis directly (as the fpingd clearly isn't doing it).

also fixed missing logmsg content for min and max for fping-controlled nodes, and improved logmessage
for why a fallback to the 'internal ping' system is being attempted.