nmis8, SUPPORT-3891, add demote_faulty_nodes config to collect long-time dead nodes more often than just once/24hrs
added config item /system/demote_faulty_nodes, default true;
updated installer to ask, for upgrades, if the user wants this to be
set to false.

updated nmis.pl to honor that option: demote to once/24hrs if not false.
if false, try every 5 min and produce a dbg-log to that effect.