nmis8, OMK-4804, ensure no more than one type=escalate op in parallel
nmis.pl now aborts any explicit type=escalate invocation that interferes
with an already running escalate (regardless of whether that running one
is part of a collect or an independent tyee=escalate).

note, however, that escalate ops as part of collect are not aborted under any
circumstances: the  process check is only made for explicit type=escalate
invocations.