nmis8, rrdfunc::getrrdashash extended for OMK-4965
now takes argument resolution, returns errors in meta response, also returns rows and rows_with_data in meta.
time and date are now the FIRST columns in the list of 'ds', not the last

also added rrdfunc::getRRDResolutions helper which returns the step and all available rras/resolutions for
a given mode.