nmis8, OMK-4866, remove badly misconfigured nodes on the fly
updated nmis::loadlocalnodetable, now offers sanitise argument.
if called with that, some crucial props are checked and dud nodes are removed
on the go, plus logmsg + warn to stderr.

nmis::upgrade_nodeconf_structure now calls loadlocalnodetable with
sanitise, because that upgrade op is the first thing that calls llnt.

updated nmis.pl, the initial upgrade-this-n-that now includes the uuid fill
ins (was done much later).